clients/team: refactor team into organization
